Improve codec, switch to bounded arrays

+++ world/utils.zig
@@ -1,5 +1,8 @@
 pub const Vec3 = struct {
-    x: f32,
-    z: f32,
-    y: f32,
+    x: u32,
+    z: u32,
+    y: u32,
+    // x: f32,
+    // z: f32,
+    // y: f32,
 };
